DAStrim declares fewer chimers, & 2 bug fixes
The logic in DAStrim that decides if a low-quality induced gap in an
A-read’s coverage is due to a chimer, adapter, or just bad sequence,
has been reworked and is now (correctly) declares fewer chimeric breaks.

The overlap pile logic was broken when initial piles are empty.

DASedit was not allocating a big enough buffer to hold the edited
   sequence (on occasion).